move UserSessionLogoutDelegate
